Passes phpstan 0.10

+++ src/Type/EntityTypeAbstractFactory.php
@@ -169,7 +169,7 @@ final class EntityTypeAbstractFactory extends AbstractAbstractFactory implements
                                                      'value' => $value,
                                                  ];
                                             } elseif (strpos($field, '_') !== false && $this->isFilter($filter)) {
-                                                $field = substr($field, 0, strrpos($field, '_'));
+                                                $field = substr($field, 0, (int)strrpos($field, '_'));
 
                                                 switch ($filter) {
                                                     case 'sort':
